Moved program list to component so render client side seperately
