+++ src/app/(frontend)/programs/page.tsx
@@ -1,20 +1,7 @@
-"use client";
-import React, { useState } from "react";
-import Link from "next/link";
-import Image from "next/image";
-import { programs_info } from "@/lib/data";
-import { CreateSlug } from "@/lib/utils";
-
-interface ProgramsInfoInterface {
-  id: number;
-  title: string;
-  description: string;
-}
+import React from "react";
+import ProgramsList from "@/components/ProgramsList";
 
 export default function ProgramsPage() {
-  // Changed from 'page' to 'ProgramsPage'
-  const [hoveredCard, setHoveredCard] = useState<string | null>(null);
-
   return (
     <main className="min-h-screen lg:container lg:mx-auto lg:py-24 p-4 py-16">
       <section className="space-y-4">
@@ -31,42 +18,7 @@ export default function ProgramsPage() {
        * End of Page Header
        */}
       <section className="grid sm:grid-cols-2 2xl:grid-cols-3 gap-8 mt-12">
-        {programs_info.map((data: ProgramsInfoInterface) => {
-          return (
-            <Link
-              key={data.id}
-              href={`/programs/${CreateSlug(data.title)}`}
-              className={`group transition-all rounded-xl border-2 border-primary-100 active:bg-primary-100 dark:border-primary-800 bg-primary-50 hover:border-primary-200 hover:bg-primary-50 dark:bg-primary-900 dark:hover:bg-primary-800 dark:active:bg-primary-700 ${
-                hoveredCard && hoveredCard !== data.title ? "opacity-50" : ""
-              }`}
-              onMouseEnter={() => setHoveredCard(data.title)}
-              onMouseLeave={() => setHoveredCard(null)}
-            >
-              <div className="w-full h-64 relative overflow-hidden">
-                <Image
-                  fill
-                  src="/media/hero-image.webp"
-                  alt={data.title}
-                  className="object-cover rounded-t-xl group-hover:contrast-125"
-                />
-              </div>
-              <div className="p-6 space-y-4">
-                <p className="text-lg font-bold leading-none tracking-tight">
-                  {data.title}
-                </p>
-                <p className="text-sm leading-5 opacity-80 line-clamp-3">
-                  {data.description}
-                </p>
-                <button
-                  type="button"
-                  className="rounded-md bg-transparent pr-2.5 py-1 text-xs font-medium capitalize text-gray-900 group-hover:underline dark:text-white dark:opacity-70"
-                >
-                  Read About The Program &rarr;
-                </button>
-              </div>
-            </Link>
-          );
-        })}
+        <ProgramsList />
       </section>
       {/*
        * End of Company Programs
